[ back / fix ] 186 : 디렉토리 구조 오류 수정
