+++ backend/src/main/java/com/example/backend/domain/analysis/service/InventoryAnalysisService.java
@@ -3,10 +3,10 @@ package com.example.backend.domain.analysis.service;
 import com.example.backend.domain.analysis.dto.CategorySummaryDTO;
 import com.example.backend.domain.analysis.dto.ItemUsageFrequencyDTO;
 import com.example.backend.domain.analysis.dto.MonthlyInventoryDTO;
-import com.example.backend.domain.Inventory.inventoryIn.entity.InventoryIn;
-import com.example.backend.domain.Inventory.inventoryIn.repository.InventoryInRepository;
-import com.example.backend.domain.Inventory.inventoryOut.entity.InventoryOut;
-import com.example.backend.domain.Inventory.inventoryOut.repository.InventoryOutRepository;
+import com.example.backend.domain.inventory.inventoryIn.entity.InventoryIn;
+import com.example.backend.domain.inventory.inventoryIn.repository.InventoryInRepository;
+import com.example.backend.domain.inventory.inventoryOut.entity.InventoryOut;
+import com.example.backend.domain.inventory.inventoryOut.repository.InventoryOutRepository;
 import com.example.backend.domain.item.entity.Item;
 import com.example.backend.domain.item.repository.ItemRepository;
 import lombok.RequiredArgsConstructor;
