+++ backend/src/main/java/com/example/backend/domain/excel/controller/ExcelExportController.java
@@ -3,8 +3,8 @@ package com.example.backend.domain.excel.controller;
 import com.example.backend.domain.excel.service.ExcelExportService;
 import com.example.backend.enums.Outbound;
 import com.example.backend.enums.Status;
-import com.example.backend.domain.Inventory.inventoryIn.service.InventoryInService;
-import com.example.backend.domain.Inventory.inventoryOut.service.InventoryOutService;
+import com.example.backend.domain.inventory.inventoryIn.service.InventoryInService;
+import com.example.backend.domain.inventory.inventoryOut.service.InventoryOutService;
 import com.example.backend.domain.item.service.ItemService;
 import com.example.backend.domain.itemInstance.dto.response.ItemInstanceResponseDto;
 import com.example.backend.domain.itemInstance.service.ItemInstanceService;
