+++ backend/src/main/java/com/example/backend/domain/item/entity/Item.java
@@ -4,8 +4,8 @@ import com.example.backend.global.auditable.Auditable;
 import com.example.backend.domain.category.entity.Category;
 import com.example.backend.domain.managementDashboard.entity.ManagementDashboard;
 import com.example.backend.enums.Status;
-import com.example.backend.domain.Inventory.inventoryIn.entity.InventoryIn;
-import com.example.backend.domain.Inventory.inventoryOut.entity.InventoryOut;
+import com.example.backend.domain.inventory.inventoryIn.entity.InventoryIn;
+import com.example.backend.domain.inventory.inventoryOut.entity.InventoryOut;
 import com.example.backend.domain.itemInstance.entity.ItemInstance;
 import com.example.backend.domain.supply.supplyRequest.entity.SupplyRequest;
 import jakarta.persistence.*;
