+++ backend/src/main/java/com/example/backend/domain/mainDashboard/repository/MainDashBoardRepository.java
@@ -8,6 +8,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
 public interface MainDashBoardRepository extends JpaRepository<MainDashboard,Long> {
-    @Query("SELECT new com.example.backend.mainDashboard.dto.MainDashBoardResponseDto(m.id,m.name,m.createdAt,m.modifiedAt) from MainDashboard m")
+    @Query("SELECT new com.example.backend.domain.mainDashboard.dto.MainDashBoardResponseDto(m.id,m.name,m.createdAt,m.modifiedAt) from MainDashboard m")
     Page<MainDashBoardResponseDto> findAllUserDtos(Pageable pageable);
 }
