+++ backend/src/main/java/com/example/backend/domain/registerItem/service/RegisterItemService.java
@@ -12,8 +12,8 @@ import com.example.backend.enums.Status;
 import com.example.backend.global.exception.BusinessLogicException;
 import com.example.backend.global.exception.ExceptionCode;
 import com.example.backend.global.utils.service.ImageService;
-import com.example.backend.domain.Inventory.inventoryIn.dto.request.InventoryInRequestDto;
-import com.example.backend.domain.Inventory.inventoryIn.service.InventoryInService;
+import com.example.backend.domain.inventory.inventoryIn.dto.request.InventoryInRequestDto;
+import com.example.backend.domain.inventory.inventoryIn.service.InventoryInService;
 import com.example.backend.domain.item.dto.request.ItemRequestDto;
 import com.example.backend.domain.item.dto.response.ItemResponseDto;
 import com.example.backend.domain.item.entity.Item;
