+++ backend/src/main/java/com/example/backend/domain/supply/supplyRequest/service/SupplyRequestService.java
@@ -13,10 +13,10 @@ import com.example.backend.enums.Outbound;
 import com.example.backend.enums.Status;
 import com.example.backend.global.exception.BusinessLogicException;
 import com.example.backend.global.exception.ExceptionCode;
-import com.example.backend.domain.Inventory.inventoryIn.dto.request.InventoryInRequestDto;
-import com.example.backend.domain.Inventory.inventoryIn.service.InventoryInService;
-import com.example.backend.domain.Inventory.inventoryOut.dto.request.InventoryOutRequestDto;
-import com.example.backend.domain.Inventory.inventoryOut.service.InventoryOutService;
+import com.example.backend.domain.inventory.inventoryIn.dto.request.InventoryInRequestDto;
+import com.example.backend.domain.inventory.inventoryIn.service.InventoryInService;
+import com.example.backend.domain.inventory.inventoryOut.dto.request.InventoryOutRequestDto;
+import com.example.backend.domain.inventory.inventoryOut.service.InventoryOutService;
 import com.example.backend.domain.item.entity.Item;
 import com.example.backend.domain.item.repository.ItemRepository;
 import com.example.backend.domain.itemInstance.dto.request.UpdateItemInstanceStatusRequestDto;
