+++ backend/src/main/java/com/example/backend/domain/supply/supplyReturn/repository/SupplyReturnRepository.java
@@ -12,12 +12,12 @@ import org.springframework.data.repository.query.Param;
 
 public interface SupplyReturnRepository extends JpaRepository<SupplyReturn, Long> {
 
-    @Query("SELECT new com.example.backend.supplyReturn.dto.response.SupplyReturnResponseDto "+
+    @Query("SELECT new com.example.backend.domain.supply.supplyReturn.dto.response.SupplyReturnResponseDto "+
             "( s.id,s.supplyRequest.id,s.user.id,s.managementDashboard.id,s.item.id,s.serialNumber,s.productName,s.quantity,s.useDate,s.returnDate,s.approvalStatus,s.createdAt,s.outbound) "+
             "from SupplyReturn s ")
     Page<SupplyReturnResponseDto> findAllSupplyReturn(Pageable pageable);
 
-    @Query("SELECT new com.example.backend.supplyReturn.dto.response.SupplyReturnResponseDto "+
+    @Query("SELECT new com.example.backend.domain.supply.supplyReturn.dto.response.SupplyReturnResponseDto "+
             "( s.id,s.supplyRequest.id,s.user.id,s.managementDashboard.id,s.item.id,s.serialNumber,s.productName,s.quantity,s.useDate,s.returnDate,s.approvalStatus,s.createdAt,s.outbound) "+
             "from SupplyReturn s " +
             "where s.approvalStatus = :approvalStatus")
