+++ backend/src/main/java/com/example/backend/domain/supply/supplyReturn/service/SupplyReturnService.java
@@ -9,8 +9,8 @@ import com.example.backend.enums.ApprovalStatus;
 import com.example.backend.enums.Inbound;
 import com.example.backend.global.exception.BusinessLogicException;
 import com.example.backend.global.exception.ExceptionCode;
-import com.example.backend.domain.Inventory.inventoryIn.dto.request.InventoryInRequestDto;
-import com.example.backend.domain.Inventory.inventoryIn.service.InventoryInService;
+import com.example.backend.domain.inventory.inventoryIn.dto.request.InventoryInRequestDto;
+import com.example.backend.domain.inventory.inventoryIn.service.InventoryInService;
 import com.example.backend.domain.item.entity.Item;
 import com.example.backend.domain.item.repository.ItemRepository;
 import com.example.backend.domain.managementDashboard.entity.ManagementDashboard;
